refactor: move channels_v2/web_channel.py to channels/web_channel.py
